+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0).
 #### Changed
 - `system` Fonts Awesome updated to version 7.3.1
 - `admin` Deep refactoring of filter functionality
+- `admin` Improved online report
 
 #### Deprecated
 - Nothing